Merge release 6 changes from Master to dev

+++ src/v2i-hub/SPaTLoggerPlugin/src/SPaTLoggerPlugin.cpp
@@ -132,148 +132,148 @@ void SPaTLoggerPlugin::HandleSpatMessage(SpatMessage &msg, routeable_message &ro
 	char *SpatOut;
         cJSON *SpatRoot, *SpatMessageContent, *_SpatMessageContent;
 	
-	PLOG(logDEBUG)<<"HandleSPaTMessage";
-	auto spat = msg.get_j2735_data();
-	
-	std::stringstream direction_hex;
-        direction_hex<<"01";
+	PLOG(logDEBUG)<<"HandleSPaTMessage" << std::endl;
+        try {
+                auto spat = msg.get_j2735_data();
+                
+                std::stringstream direction_hex;
+                direction_hex<<"01";
 
-        std::stringstream signStatus_hex;
-        signStatus_hex<<"00";
+                std::stringstream signStatus_hex;
+                signStatus_hex<<"00";
 
-	uint16_t intersectionId = spat->intersections.list.array[0]->id.id;
-        uint16_t intersectionId_sw = __builtin_bswap16(abs(intersectionId));
-	std::stringstream intersectionId_size;
-	intersectionId_size << intersectionId;
-	std::stringstream intersectionId_int_hex;
-        if (intersectionId_size.str().length() == 1)
-	{
-		intersectionId_int_hex<<"0"<<std::hex<<intersectionId_sw;
-	        unsigned int intersectionId_size2 = intersectionId_int_hex.str().size();
-		std::stringstream intersectionId_hex;
-        	intersectionId_hex<<intersectionId_int_hex.str()[1]<<' '<<intersectionId_int_hex.str()[2]<<intersectionId_int_hex.str()[3];
-	}
-	else
-		intersectionId_int_hex<<std::hex<<intersectionId_sw;
+                // Retrieve spat intersectionId uint16_t       
+                uint16_t intersectionId = spat->intersections.list.array[0]->id.id;
+                // Swap bytes to little endian
+                uint16_t intersectionId_sw = __builtin_bswap16(intersectionId);
+                std::stringstream intersectionId_int_hex;
+                // Creates 2 byte hex string including leading zeros
+                intersectionId_int_hex<< std::setfill('0') << std::setw(sizeof(uint16_t)*2) << std::hex <<intersectionId_sw;
                 std::stringstream intersectionId_hex;
+                // Format string include spaces between each byte 
                 intersectionId_hex<<intersectionId_int_hex.str()[0]<<intersectionId_int_hex.str()[1]<<' '<<intersectionId_int_hex.str()[2]<<intersectionId_int_hex.str()[3];
-	
-	std::string interstatus = std::bitset<8>(spat->intersections.list.array[0]->status.buf).to_string();
-	int interstatint_sw = stoi(interstatus);
-        std::stringstream interstatint_size;
-        interstatint_size << interstatint_sw;
-        std::stringstream interstatint_int_hex;
-        if (interstatint_size.str().length() == 1)
-        {
-                interstatint_int_hex<<"0"<<std::hex<<interstatint_sw;
-                unsigned int interstatint_size2 = interstatint_int_hex.str().size();
-                std::stringstream interstatint_hex;
-                interstatint_hex<<interstatint_int_hex.str()[1];
-        }
-        else
-                interstatint_int_hex<<std::hex<<interstatint_sw;
+
+                // Retrieve spat interstatus int8_t       
+                std::string interstatus = std::bitset<8>(spat->intersections.list.array[0]->status.buf).to_string();
+                int interstatint_sw = stoi(interstatus);
+                std::stringstream interstatint_int_hex;
+                // Creates 1 byte hex string including leading zeros
+                interstatint_int_hex<< std::setfill('0') << std::setw(sizeof(int8_t)*2) << std::hex << interstatint_sw;
                 std::stringstream interstatint_hex;
                 interstatint_hex<<interstatint_int_hex.str()[0]<<interstatint_int_hex.str()[1];
 
-        uint32_t spatreceivetime = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-        uint32_t spatreceivetime_sw = __builtin_bswap32(spatreceivetime);
-        std::stringstream spatreceivetime_int_hex;
-        spatreceivetime_int_hex<<std::hex<<spatreceivetime_sw;
-        unsigned int spatreceivetime_size = spatreceivetime_int_hex.str().size();
-        std::stringstream spatreceivetime_hex;
-        spatreceivetime_hex<<spatreceivetime_int_hex.str()[0]<<spatreceivetime_int_hex.str()[1];
-        for (int fsize = 2; fsize < spatreceivetime_size; fsize = fsize+2) {
-        spatreceivetime_hex << ' ' << spatreceivetime_int_hex.str()[fsize] << spatreceivetime_int_hex.str()[fsize+1];
+                // Retrieve spat receivetime uint32_t       
+                uint32_t spatreceivetime = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+                // Swap bytes to little endian
+                uint32_t spatreceivetime_sw = __builtin_bswap32(spatreceivetime);
+                std::stringstream spatreceivetime_int_hex;
+                // Creates 4 byte hex string including leading zeros
+                spatreceivetime_int_hex<<std::setfill('0') << std::setw(sizeof(uint32_t)*2) << std::hex <<spatreceivetime_sw;
+                std::stringstream spatreceivetime_hex;
+                // Format string include spaces between each byte 
+                spatreceivetime_hex<<spatreceivetime_int_hex.str()[0]<<spatreceivetime_int_hex.str()[1] << ' '
+                        <<spatreceivetime_int_hex.str()[2]<<spatreceivetime_int_hex.str()[3] << ' '
+                        <<spatreceivetime_int_hex.str()[4]<<spatreceivetime_int_hex.str()[5] << ' '
+                        <<spatreceivetime_int_hex.str()[6]<<spatreceivetime_int_hex.str()[7];
+        
+                // Retrieve spat receivetimemillis uint16_t       
+                uint64_t spatreceivetimemillis = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+                unsigned int spatmillis = spatreceivetimemillis - (spatreceivetime*1000);
+                uint16_t spatmillis16 = (uint16_t) spatmillis;
+                // Swap bytes to little endian
+                uint16_t spatmillis16_sw = __builtin_bswap16(spatmillis16);
+                std::stringstream spatmillis16_int_hex;
+                // Creates 2 byte hex string including leading zeros
+                spatmillis16_int_hex<< std::setfill('0') << std::setw(sizeof(uint16_t)*2) << std::hex <<spatmillis;
+                std::stringstream spatmillis16_hex;
+                // Format string include spaces between each byte
+                spatmillis16_hex<< spatmillis16_int_hex.str()[0]<<spatmillis16_int_hex.str()[1]<<' '<<spatmillis16_int_hex.str()[2]<<spatmillis16_int_hex.str()[3];
+
+                std::stringstream IsCertPresent_hex;
+                IsCertPresent_hex<<"00";
+
+                // Retrieve spat spatsize uint16_t       
+                uint16_t spat_size;
+                if (routeableMsg.get_payload_str().length()%4 == 0){
+                        spat_size = routeableMsg.get_payload_str().length()/2;
+                }
+                else{
+                        spat_size = (routeableMsg.get_payload_str().length()/2)+1;
+                }
+                // Swap bytes to little endian
+                uint16_t spat_size_test = __builtin_bswap16(spat_size);
+                std::stringstream spat_size_int_hex;
+                // Creates 2 byte hex string including leading zeros
+                spat_size_int_hex<< std::setfill('0') << std::setw(sizeof(uint16_t)*2) << std::hex << spat_size_test;
+                std::stringstream spat_size_hex;
+                // Format string include spaces between each byte
+                spat_size_hex<<spat_size_int_hex.str()[0]<<spat_size_int_hex.str()[1]<<' '<<spat_size_int_hex.str()[2]<<spat_size_int_hex.str()[3];
+                
+                
+                SpatRoot = cJSON_CreateObject(); // create root node
+                SpatMessageContent = cJSON_CreateArray(); // create root array
+
+                cJSON_AddItemToObject(SpatRoot, "SpatMessageContent", SpatMessageContent); // add SpatMessageContent array to Spatroot
+
+                cJSON_AddItemToArray(SpatMessageContent, _SpatMessageContent = cJSON_CreateObject()); //add message content to SpatMessageContent array
+                cJSON_AddItemToObject(_SpatMessageContent, "IntersectionID", cJSON_CreateNumber(intersectionId)); // Intersection_ID
+                cJSON_AddItemToObject(_SpatMessageContent, "IntersecionStatus", cJSON_CreateNumber(interstatint_sw)); // Intersection_status
+                cJSON_AddItemToObject(_SpatMessageContent, "payload", cJSON_CreateString(routeableMsg.get_payload_str().c_str())); // payload
+
+
+                std::stringstream metadata;
+                std::string spat_output_payload = routeableMsg.get_payload_str();
+
+                // ----------------- Header Description-------------------------
+                // uint8_t direction ("01")
+                // int16_t intersectionId
+                // int8_t intersectionStatus
+                // uint32_t spatreceivetime
+                // uint16_t spatmillis
+                // int8_t signStatus
+                // int8_t isCertPresent
+                // uint16_t length
+                // -------------------------------------------------------------
+                metadata<<direction_hex.str()<<' '<<intersectionId_hex.str()<<' '<<interstatint_hex.str()<<' '<<spatreceivetime_hex.str()<<' '<<spatmillis16_hex.str()<<' '<<signStatus_hex.str()<<' '<<IsCertPresent_hex.str()<<' '<<spat_size_hex.str();
+                
+                unsigned int payload_size = spat_output_payload.size();
+                std::stringstream spat_output_data;
+                spat_output_data<<metadata.str();
+                for (int bsize = 0; bsize <= payload_size; bsize = bsize+2) {
+                spat_output_data << ' ' << spat_output_payload[bsize] << spat_output_payload[bsize+1];
+                }
+                int actualsize = spat_output_data.str().size();
+                int showmethesize = ((spat_output_data.str().size()+1)/3);
+                unsigned int binary_array[((spat_output_data.str().size()+1)/3)];
+                int dsize = 0;
+                while (spat_output_data.good() && dsize < spat_output_data.str().size()){
+                        spat_output_data >> std::hex >> binary_array[dsize];
+                        ++dsize;
+                }
+                unsigned char binary_output[((spat_output_data.str().size()+1)/3)];
+                for(int k=0;k<(spat_output_data.str().size()+1)/3;k++) {
+                binary_output[k]=static_cast<char>(binary_array[k]);
+                }
+                std::ofstream _logFileBin(_curFilenamebin, std::ios::out | std::ios::binary | std::ios::app);
+                _logFileBin.write((const char*)binary_output,sizeof(binary_output));
+
+                PLOG(logDEBUG)<<"Logging SPaTMessage data";
+                SpatOut = cJSON_Print(SpatRoot);
+                _logFile << SpatOut;
+                free(SpatOut);
         }
-
-        uint64_t spatreceivetimemillis = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-        unsigned int spatmillis = spatreceivetimemillis - (spatreceivetime*1000);
-        uint16_t spatmillis16 = (uint16_t) spatmillis;
-        uint16_t spatmillis16_sw = __builtin_bswap16(abs(spatmillis16));
-        std::stringstream spatmillis16_int_hex;
-        spatmillis16_int_hex<<"0"<<std::hex<<spatmillis;
-        std::stringstream spatmillis16_size;
-        spatmillis16_size<<spatmillis16_int_hex.str()[2]<<spatmillis16_int_hex.str()[3];
-        std::stringstream spatmillis16_hex;
-        if (spatmillis16_int_hex.str()[2] == NULL){
-                spatmillis16_hex<<"0"<<spatmillis16_int_hex.str()[3]<<' '<<spatmillis16_int_hex.str()[0]<<spatmillis16_int_hex.str()[1];
+        catch (J2735Exception &e) {
+           FILE_LOG(logERROR) << "Exception caught " << e.what() << std::endl << e.GetBacktrace() << std::endl;
         }
-        else if (spatmillis16_int_hex.str()[3] == NULL){
-                spatmillis16_hex<<spatmillis16_int_hex.str()[2]<<"0"<<' '<<spatmillis16_int_hex.str()[0]<<spatmillis16_int_hex.str()[1];
-        }
-
-        else{
-                spatmillis16_hex<<spatmillis16_int_hex.str()[2]<<spatmillis16_int_hex.str()[3]<<' '<<spatmillis16_int_hex.str()[0]<<spatmillis16_int_hex.str()[1];
-        }
-
-	//Add is_cert_present
-        std::stringstream IsCertPresent_hex;
-        IsCertPresent_hex<<"00";
-
-	uint16_t spat_size;
-        if (routeableMsg.get_payload_str().length()%4 == 0){
-                spat_size = routeableMsg.get_payload_str().length()/2;
-        }
-        else{
-                spat_size = (routeableMsg.get_payload_str().length()/2)+1;
-        }
-        uint16_t spat_size_test = __builtin_bswap16(spat_size);
-        std::stringstream spat_size_int_hex;
-        spat_size_int_hex<<std::hex<< spat_size_test;
-	std::stringstream spat_size_hex;
-        spat_size_hex<<spat_size_int_hex.str()[0]<<spat_size_int_hex.str()[1]<<' '<<spat_size_int_hex.str()[2]<<spat_size_int_hex.str()[3];
-	
-	
-	SpatRoot = cJSON_CreateObject(); // create root node
-        SpatMessageContent = cJSON_CreateArray(); // create root array
-
-        cJSON_AddItemToObject(SpatRoot, "SpatMessageContent", SpatMessageContent); // add SpatMessageContent array to Spatroot
-
-        cJSON_AddItemToArray(SpatMessageContent, _SpatMessageContent = cJSON_CreateObject()); //add message content to SpatMessageContent array
-        cJSON_AddItemToObject(_SpatMessageContent, "IntersectionID", cJSON_CreateNumber(intersectionId)); // Intersection_ID
-        cJSON_AddItemToObject(_SpatMessageContent, "IntersecionStatus", cJSON_CreateNumber(interstatint_sw)); // Intersection_status
-	cJSON_AddItemToObject(_SpatMessageContent, "payload", cJSON_CreateString(routeableMsg.get_payload_str().c_str())); // payload
-
-
-	std::stringstream metadata;
-        std::string spat_output_payload = routeableMsg.get_payload_str();
-
-        metadata<<direction_hex.str()<<' '<<intersectionId_hex.str()<<' '<<interstatint_hex.str()<<' '<<spatreceivetime_hex.str()<<' '<<spatmillis16_hex.str()<<' '<<signStatus_hex.str()<<' '<<IsCertPresent_hex.str()<<' '<<spat_size_hex.str();
-
-        unsigned int payload_size = spat_output_payload.size();
-        std::stringstream spat_output_data;
-        spat_output_data<<metadata.str();
-        for (int bsize = 0; bsize <= payload_size; bsize = bsize+2) {
-        spat_output_data << ' ' << spat_output_payload[bsize] << spat_output_payload[bsize+1];
-        }
-        int actualsize = spat_output_data.str().size();
-        int showmethesize = ((spat_output_data.str().size()+1)/3);
-        unsigned int binary_array[((spat_output_data.str().size()+1)/3)];
-        int dsize = 0;
-        while (spat_output_data.good() && dsize < spat_output_data.str().size()){
-                spat_output_data >> std::hex >> binary_array[dsize];
-                ++dsize;
-        }
-        unsigned char binary_output[((spat_output_data.str().size()+1)/3)];
-        for(int k=0;k<(spat_output_data.str().size()+1)/3;k++) {
-        binary_output[k]=static_cast<char>(binary_array[k]);
-        }
-        std::ofstream _logFileBin(_curFilenamebin, std::ios::out | std::ios::binary | std::ios::app);
-        _logFileBin.write((const char*)binary_output,sizeof(binary_output));
-
-	PLOG(logDEBUG)<<"Logging SPaTMessage data";
-	SpatOut = cJSON_Print(SpatRoot);
-        _logFile << SpatOut;
-        free(SpatOut);
 
 }
 
 
 
 /**
- *  Opens a new log file in the directory specified of specified name for logging SPaT messages and
- *  inserts a header row with names of fields that will be logged when data is received. If a log file
- *  with the same name already exists before opening a new file, it's renamed with current timestamp suffix.
+ * Opens a new log file in the directory specified of specified name for logging SPaT messages. Once the
+ * current binary logfile size reaches the configurable maxSize this file is closed, renamed by the current
+ * time and date and moved to a /ode/ directory where it can be sent to an ODE using the filewatchscript.sh.
  */
 void SPaTLoggerPlugin::OpenSPaTLogFile()
 {
@@ -288,14 +288,17 @@ void SPaTLoggerPlugin::OpenSPaTLogFile()
         std::string newbinFilename = _fileDirectory + "/ode/" + _filename + GetCurDateTimeStr() + ".bin";
 
         int error;
-        error = std::rename(_curFilename.c_str(), newFilename.c_str() );
-        if ( error != 0 ) {
-                FILE_LOG(logERROR) << "Failed to mv " << _curFilename.c_str() << " to " << newFilename.c_str() << std::endl;
-        }
-        error = std::rename(_curFilenamebin.c_str(), newbinFilename.c_str() );
-        if ( error != 0 ) {
-                FILE_LOG(logERROR) << "Failed to mv " << _curFilenamebin.c_str() << " to " << newbinFilename.c_str() << std::endl;
-        }
+        if ( boost::filesystem::exists( _curFilenamebin.c_str() ) ) {
+                error = std::rename(_curFilename.c_str(), newFilename.c_str() );
+    	        if ( error != 0 ) {
+        	        FILE_LOG(logERROR) << "Failed to mv " << _curFilename.c_str() << " to " << newFilename.c_str() << std::endl;
+    	        }
+	
+    	        error = std::rename(_curFilenamebin.c_str(), newbinFilename.c_str() );
+                if ( error != 0 ) {
+                        FILE_LOG(logERROR) << "Failed to mv " << _curFilenamebin.c_str() << " to " << newbinFilename.c_str() << std::endl;
+                }
+	}
         _logFile.open(_curFilename);
         _logFilebin.open(_curFilenamebin, std::ios::out | std::ios::binary | std::ios::app);
 	if (!_logFile.is_open())
